refactor incoming order feature
