+++ packages/flipper_dashboard/lib/features/incoming_orders/providers/incoming_orders_provider.dart
@@ -1,0 +1,10 @@
+import 'package:flipper_models/realm_model_export.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:flipper_services/proxy.dart';
+
+// Provider to cache transaction items for each request
+final transactionItemsProvider =
+    FutureProvider.family<List<TransactionItem>, String>(
+  (ref, requestId) =>
+      ProxyService.strategy.transactionItems(requestId: requestId),
+);
