+++ packages/flipper_dashboard/lib/features/incoming_orders/widgets/status_delivery_info.dart
@@ -1,0 +1,192 @@
+import 'package:flipper_models/realm_model_export.dart';
+import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
+
+class StatusDeliveryInfo extends StatelessWidget {
+  final InventoryRequest request;
+
+  const StatusDeliveryInfo({
+    Key? key,
+    required this.request,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          'Status & Delivery',
+          style: TextStyle(
+            fontSize: 16,
+            fontWeight: FontWeight.w600,
+            color: Colors.grey[800],
+          ),
+        ),
+        SizedBox(height: 12),
+        Container(
+          padding: EdgeInsets.all(16),
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.circular(12),
+            border: Border.all(color: Colors.grey[200]!),
+          ),
+          child: Column(
+            children: [
+              _buildStatusInfo(),
+              SizedBox(height: 12),
+              _buildDeliveryInfo(),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget _buildStatusInfo() {
+    return Row(
+      children: [
+        Container(
+          padding: EdgeInsets.all(8),
+          decoration: BoxDecoration(
+            color: _getStatusColor(request.status).withOpacity(0.1),
+            borderRadius: BorderRadius.circular(8),
+          ),
+          child: Icon(
+            _getStatusIcon(request.status),
+            color: _getStatusColor(request.status),
+            size: 20,
+          ),
+        ),
+        SizedBox(width: 12),
+        Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(
+              'Status',
+              style: TextStyle(
+                fontSize: 14,
+                color: Colors.grey[600],
+                fontWeight: FontWeight.w500,
+              ),
+            ),
+            Text(
+              request.status?.toUpperCase() ?? 'N/A',
+              style: TextStyle(
+                fontSize: 14,
+                color: _getStatusColor(request.status),
+                fontWeight: FontWeight.w600,
+              ),
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+
+  Widget _buildDeliveryInfo() {
+    return Row(
+      children: [
+        Container(
+          padding: EdgeInsets.all(8),
+          decoration: BoxDecoration(
+            color: Colors.purple[50],
+            borderRadius: BorderRadius.circular(8),
+          ),
+          child: Icon(Icons.calendar_today, color: Colors.purple[700], size: 20),
+        ),
+        SizedBox(width: 12),
+        Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(
+              'Requested On',
+              style: TextStyle(
+                fontSize: 14,
+                color: Colors.grey[600],
+                fontWeight: FontWeight.w500,
+              ),
+            ),
+            Text(
+              DateFormat('MMM dd, yyyy HH:mm').format(request.createdAt!),
+              style: TextStyle(
+                fontSize: 14,
+                color: Colors.purple[700],
+                fontWeight: FontWeight.w600,
+              ),
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+
+  Color _getStatusColor(String? status) {
+    switch (status?.toLowerCase()) {
+      case 'pending':
+        return Colors.orange[700]!;
+      case 'approved':
+        return Colors.green[700]!;
+      case 'voided':
+        return Colors.red[700]!;
+      default:
+        return Colors.grey[700]!;
+    }
+  }
+
+  IconData _getStatusIcon(String? status) {
+    switch (status?.toLowerCase()) {
+      case 'pending':
+        return Icons.pending;
+      case 'approved':
+        return Icons.check_circle;
+      case 'voided':
+        return Icons.cancel;
+      default:
+        return Icons.help;
+    }
+  }
+}
+
+class OrderNote extends StatelessWidget {
+  final InventoryRequest request;
+
+  const OrderNote({
+    Key? key,
+    required this.request,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          'Order Note',
+          style: TextStyle(
+            fontSize: 16,
+            fontWeight: FontWeight.w600,
+            color: Colors.grey[800],
+          ),
+        ),
+        SizedBox(height: 8),
+        Container(
+          width: double.infinity,
+          padding: EdgeInsets.all(16),
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.circular(12),
+            border: Border.all(color: Colors.grey[200]!),
+          ),
+          child: Text(
+            request.orderNote ?? '',
+            style: TextStyle(
+              fontSize: 14,
+              color: Colors.grey[700],
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
